Split out file utils into separate submodule
